refactor(engine): extract decision evaluation logic
* Extract the Decision evaluation logic from the BpmnDecisionBehavior class into a new DecisionBehavior class to make it reusable. An instance of the new DecisionBehavior class can be obtained from the BpmnDecisionBehavior instance, by calling BpmnDecisionBehavior#getDecisionBehavior(). By providing a getter, we can reuse the already created DecisionBehavior instance. This pattern is also used for the CatchEventBehavior.
