+++ engine/src/main/java/io/camunda/zeebe/engine/processing/bpmn/behavior/BpmnBehaviors.java
@@ -18,7 +18,7 @@ public interface BpmnBehaviors {
 
   ExpressionProcessor expressionBehavior();
 
-  BpmnDecisionBehavior decisionBehavior();
+  BpmnDecisionBehavior bpmnDecisionBehavior();
 
   BpmnVariableMappingBehavior variableMappingBehavior();
